web_thread

Trait Post

pub trait Post: AsJs {
    // Provided method
    fn transferables(&self) -> Array { ... }
}

Objects that can be sent via postMessage. A type that is Post supports being serialized into a JavaScript object that can be sent using postMessage, and also getting an array of subobjects that must be transferred.

Provided Methods

fn transferables(&self) -> Array

Get a list of the objects that must be transferred when calling postMessage.

The default implementation returns an empty array.